Contextual keyboard enter search implemented.
When located in a search field and keyboard enter key is pressed 
the corresponding search function is called.
This behaviour is achieved thanks to both hb-enter directive and 
new controller defined navigation functions directing to either 
a result list or an object card view depending on the number of 
objects resulting from end user selection.